+++ crates/captval/src/lib.rs
@@ -279,24 +279,32 @@
 //!
 //! This version of captval requires Rust v1.82 or later.
 
-// mod captcha;
+mod captcha;
 // mod client;
 mod captval;
 mod error;
-// mod request;
 mod hcaptcha;
 mod remoteip;
+// mod request;
 mod response;
 
-// pub use captcha::Captcha;
+pub use captcha::Captcha;
 // pub use client::Client;
 // pub use client::VERIFY_URL;
 pub use error::Error;
-pub use hcaptcha::Code;
 // pub use request::Request;
 pub use response::Response;
 
-// pub(crate) use remoteip::Remoteip;
+pub(crate) use remoteip::Remoteip;
 
 pub use crate::captval::Captval;
 // pub use captval_derive::*;
+
+#[cfg(feature = "hcaptcha")]
+pub use hcaptcha::ClientResponse;
+#[cfg(feature = "hcaptcha")]
+pub use hcaptcha::Code;
+// #[cfg(feature = "hcaptcha")]
+// pub(crate) use hcaptcha::Secret;
+#[cfg(feature = "hcaptcha")]
+pub(crate) use hcaptcha::Sitekey;